frontend/requirements.txt
streamlit
requests
pandas
prometheus-client
psutil
pyyaml